Sync. w. JSBSIm one more time to fix at least one bug

+++ src/FDM/JSBSim/models/FGPropulsion.cpp
@@ -216,7 +216,7 @@ void FGPropulsion::InitRunning(int n)
 {
   if (n > 0) { // A specific engine is supposed to be initialized
 
-    if (n >= GetNumEngines() ) {
+    if (n >= (int)GetNumEngines() ) {
       cerr << "Tried to initialize a non-existent engine!" << endl;
       throw;
     }
